Create motions correctly
