+++ webots_simulation/controllers/robot_controller/motion_functions.py
@@ -232,3 +232,6 @@ def move(robot, sync, position, joint, rotation='', side='', delay=250, accelera
             robot.motor_set_position_sync(robot.motors[motor], robot.sensors[motor], position, delay)
     except:
         raise Exception(f'Wrong args. Your args resulted in trying to move "{motor}"...')
+
+
+
