+++ webots_simulation/controllers/robot_controller/robot_controller.py
@@ -8,6 +8,7 @@ import json
 import yaml
 import Levenshtein
 import motion_functions
+import helpers
 from controller import Robot, Motor
 
 # import supercontroller
@@ -34,7 +35,7 @@ class NaoMotorController(SuperController):
         if 'instruction' in body:
             self.instruction = self.findMatchingMotionFunction(body['instruction'])
         else:
-            self.createNewMotion(body)
+            helpers.createNewMotion(body)
 
     def findMatchingMotionFunction(self, s):
         '''
@@ -47,12 +48,7 @@ class NaoMotorController(SuperController):
         RETURNS:
             The name of the closest matching motion-function
         '''
-
-        def getMotionFunctions():
-            '''Returns a list of all motion-functions available in motion_functions.py'''
-            return [motion_function for motion_function in dir(motion_functions) if not motion_function.startswith('__')]
-
-        func_list = getMotionFunctions()
+        func_list = helpers.getMotionFunctions()
         min_edit_distance = 5 # The maximum distance allowed
         closest_match = None
         for func in func_list:
@@ -61,68 +57,6 @@ class NaoMotorController(SuperController):
                 min_edit_distance = edit_distance
                 closest_match = func
         return closest_match
-
-        def createNewMotion(body):
-            '''Creates a new motion function in motion_functions.py from an RML JSON body'''
-
-            # Check if name already exists
-            existing = getMotionFunctions()
-            if body['def'] in existing:
-                print("A motion function with the same name already exist, please change the name")
-                return
-
-            # Open file
-            with open("motion_functions.py" as f):
-                # Write to file
-                f.write('\n')
-
-                # Define function
-                f.write(f"def {body['def']}:\n")
-
-                # Write function content
-                for cmd in body['commands']:
-                    if 'repeat' in cmd:
-                        writeRepeat(cmd['repeat'], cmd['amount'])
-                    else if 'multimove' in cmd:
-                        writeNewKeyframe()
-                        writeMultimove(cmd['multimove'])
-                    else if 'move' in cmd:
-                        writeNewKeyframe()
-                        writeMove(cmd['move'], True)
-                    else:
-                        print(f"unknown command {cmd}")
-
-                ## Functions for writing to the file
-                def writeRepeat(repeat, amount):
-                    for i in range(amount):
-                        writeNewKeyframe()
-                        for cmd_to_repeat in repeat:
-                            if 'multimove' in cmd_to_repeat:
-                                writeMultimove(cmd_to_repeat['multimove'])
-                            else if 'move' in cmd_to_repeat:
-                                writeMove(cmd_to_repeat['move'], True)
-                            else:
-                                print(f"unknown command {cmd_to_repeat}")
-
-                def writeMultimove(multi):
-                    for i in range(len(multi)):
-                        if i+1 == len(multi):
-                            writeMove(multi[i], True)
-                        else:
-                            writeMove(multi[i], False)
-                
-                def writeMove(move, sync):
-                    pos = move['position']
-                    joint = move['joint'].capitalize()
-                    side = move['side'][0].capitalize() if 'side' in move else ''
-                    rot = move['rotation'] if 'rotation' in move else ''
-                    f.write(f"\tmove(robot, {sync}, {pos}, {joint}, {rot}, {side})\n")
-
-                def writeNewKeyframe():
-                    f.write("\t# Keyframe\n")
-
-
-            
 
     # Controller loop
     def run(self):
