+++ webots_simulation/worlds/.example.wbproj
@@ -1,5 +1,5 @@
 Webots Project File version R2022b
-perspectives: 000000ff00000000fd0000000200000001000001db00000457fc0200000001fb0000001400540065007800740045006400690074006f00720100000016000004570000004100ffffff0000000300000a00000000ddfc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c010000000000000a000000006900ffffff000008230000045700000001000000020000000100000008fc00000000
+perspectives: 000000ff00000000fd0000000200000001000001db0000037bfc0200000001fb0000001400540065007800740045006400690074006f007201000000160000037b0000004100ffffff0000000300000a00000001b9fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c010000000000000a000000006900ffffff000008230000037b00000001000000020000000100000008fc00000000
 simulationViewPerspectives: 000000ff000000010000000200000100000004720100000002010000000100
 sceneTreePerspectives: 000000ff00000001000000030000001e00000215000000fa0100000002010000000200
 maximizedDockId: -1
